Project1/project2/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-tinymce==3.6.1
numpy==1.26.4
scipy==1.12.0
sqlparse==0.4.4
tzdata==2023.4
